pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "gaql"
version = "1.1.0"
description = "A command line interface to the Google Ads Query Language (GAQL). Run with `gaql`"

authors = [
  "Ben Ryves <bryves@gmail.com>"
]

readme = 'README.md'
repository = 'https://github.com/getyourguide/gaql-cli'
homepage = 'https://github.com/getyourguide/gaql-cli'

keywords = ['gaql', 'google-ads', 'cli']

[tool.poetry.scripts]
gaql = 'gaql.client.commands:cli'

[tool.poetry.dependencies]
python = "^3.7"
toml = "^0.10.1"

google-ads={ version = "8.0.0" }
click={ version = "^7.1.2" }
prompt_toolkit={ version = "^3.0.8" }
pygments={ version = "^2.7.3" }

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
